# docs/设计/接口契约/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Campus-Link API",
    "description": "重庆工程学院计算机专业学生交流论坛（MVP）。错误码分段见技术方案第 5 节。",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/notifications/read-all": {
      "put": {
        "tags": [
          "notification"
        ],
        "summary": "全部标记已读（需登录）：updated 为本次新标记的条数，无未读时为 0",
        "operationId": "markAllRead",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMarkAllReadResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录（4001）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "系统繁忙，请稍后再试（9999）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notification"
        ],
        "summary": "我的通知（需登录）：unread 缺省不限、true 只未读、false 只已读；按时间倒序分页 page/size",
        "operationId": "list_2",
        "parameters": [
          {
            "name": "unread",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageVoNotificationVo"
                }
              }
            }
          },
          "400": {
            "description": "参数错误（1001）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "未登录（4001）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "系统繁忙，请稍后再试（9999）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/notifications/unread-count": {
      "get": {
        "tags": [
          "notification"
        ],
        "summary": "未读通知数（需登录）：顶栏铃铛角标",
        "operationId": "unreadCount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUnreadCountResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录（4001）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "系统繁忙，请稍后再试（9999）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseMarkAllReadResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/MarkAllReadResult"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "MarkAllReadResult": {
        "type": "object",
        "properties": {
          "updated": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ApiResponsePageVoNotificationVo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageVoNotificationVo"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "NotificationVo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "actorId": {
            "type": "integer",
            "format": "int64"
          },
          "actorNickname": {
            "type": "string"
          },
          "targetType": {
            "type": "string"
          },
          "targetId": {
            "type": "integer",
            "format": "int64"
          },
          "postId": {
            "type": "integer",
            "format": "int64"
          },
          "postTitle": {
            "type": "string"
          },
          "floorNo": {
            "type": "integer",
            "format": "int32"
          },
          "read": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PageVoNotificationVo": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationVo"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseUnreadCountResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UnreadCountResult"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "UnreadCountResult": {
        "type": "object",
        "properties": {
          "unreadCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ApiError": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "登录返回的 JWT，请求头 `Authorization: Bearer <token>`。框架在进入业务代码前按端点注解裁决：标 `@SecurityRequirement` 的端点要求已登录，标 `@PublicEndpoint` 的放行，module 端点两者都没标即 fail-closed；未登录由过滤器链直接产出 401 / 4001（CR-031，SecurityConfig 已无 permitAll 兜底）。角色与资源级授权仍归业务代码的 CurrentUser（requireRole → 403 / 4002）；受保护端点必须声明本 security 并真的调用 CurrentUser，该纪律由 ArchitectureGuardTest 机器校验。",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}